requirements.txt
google-adk==1.0.0
litellm==1.15.0
pytest==6.2
